@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-  compileSdkVersion 25
-  buildToolsVersion "25.0.2"
+  compileSdkVersion 32
   defaultConfig {
     applicationId "com.xiaopo.flying.stickerview"
     minSdkVersion 14
-    targetSdkVersion 25
+    targetSdkVersion 32
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
   }
   buildTypes {
     release {
@@ -20,12 +19,12 @@ android {
 }
 
 dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-  androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+  implementation fileTree(dir: 'libs', include: ['*.jar'])
+  androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
     exclude group: 'com.android.support', module: 'support-annotations'
   })
-  compile "com.android.support:appcompat-v7:$rootProject.ext.supportLibraryVersion"
-  compile "com.github.chrisbanes:PhotoView:$rootProject.ext.photoViewVersion"
-  testCompile 'junit:junit:4.12'
-  compile project(path: ':sticker')
+  implementation "androidx.appcompat:appcompat:$rootProject.ext.appCompatVersion"
+  implementation "com.github.chrisbanes:PhotoView:$rootProject.ext.photoViewVersion"
+  testImplementation 'junit:junit:4.12'
+  implementation project(path: ':sticker')
 }
